refactor local orders state
