+++ src/lib/abacus/dydxChainTransactions.ts
@@ -40,8 +40,8 @@ import { DydxAddress } from '@/constants/wallets';
 import { type RootStore } from '@/state/_store';
 // TODO Fix cycle
 // eslint-disable-next-line import/no-cycle
-import { placeOrderTimeout } from '@/state/account';
 import { setInitializationError } from '@/state/app';
+import { placeOrderTimeout } from '@/state/localOrders';
 
 import { dd } from '../analytics/datadog';
 import { signComplianceSignature, signComplianceSignatureKeplr } from '../compliance';
