+++ src/state/account.ts
@@ -1,5 +1,4 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
-import _ from 'lodash';
 
 import {
   type AccountBalance,
@@ -21,20 +20,10 @@ import {
   type Wallet,
 } from '@/constants/abacus';
 import { OnboardingGuard, OnboardingState } from '@/constants/account';
-import { DEFAULT_SOMETHING_WENT_WRONG_ERROR_PARAMS, ErrorParams } from '@/constants/errors';
 import { LocalStorageKey } from '@/constants/localStorage';
-import {
-  CANCEL_ALL_ORDERS_KEY,
-  CancelOrderStatuses,
-  LocalCancelAllData,
-  PlaceOrderStatuses,
-  type LocalCancelOrderData,
-  type LocalPlaceOrderData,
-  type TradeTypes,
-} from '@/constants/trade';
 
 import { getLocalStorage } from '@/lib/localStorage';
-import { isOrderStatusCanceled, isOrderStatusClearable } from '@/lib/orders';
+import { isOrderStatusClearable } from '@/lib/orders';
 
 export type AccountState = {
   balances?: Record<string, AccountBalance>;
@@ -66,16 +55,10 @@ export type AccountState = {
 
   onboardingGuards: Record<OnboardingGuard, boolean | undefined>;
   onboardingState: OnboardingState;
-
   clearedOrderIds?: string[];
   unseenFillsCountPerMarket: Record<string, number>;
   hasUnseenOrderUpdates: boolean;
-  latestOrder?: Nullable<SubaccountOrder>;
   historicalPnlPeriod?: HistoricalPnlPeriods;
-  uncommittedOrderClientIds: string[];
-  localPlaceOrders: LocalPlaceOrderData[];
-  localCancelOrders: LocalCancelOrderData[];
-  localCancelAlls: Record<string, LocalCancelAllData>;
 
   restriction?: Nullable<UsageRestriction>;
   compliance?: Compliance;
@@ -110,12 +93,8 @@ const initialState: AccountState = {
   clearedOrderIds: undefined,
   unseenFillsCountPerMarket: {},
   hasUnseenOrderUpdates: false,
-  latestOrder: undefined,
-  uncommittedOrderClientIds: [],
+
   historicalPnlPeriod: undefined,
-  localPlaceOrders: [],
-  localCancelOrders: [],
-  localCancelAlls: {},
 
   // Restriction
   restriction: undefined,
@@ -144,27 +123,10 @@ export const accountSlice = createSlice({
           });
       }
 
-      const filledOrderIds = (action.payload ?? []).map((fill: SubaccountFill) => fill.orderId);
-
       return {
         ...state,
         fills: action.payload,
         unseenFillsCountPerMarket: newUnseenFillsCountPerMarket,
-        localPlaceOrders: hasNewFillUpdates
-          ? state.localPlaceOrders.map((order) =>
-              order.submissionStatus < PlaceOrderStatuses.Filled &&
-              order.orderId &&
-              filledOrderIds.includes(order.orderId)
-                ? {
-                    ...order,
-                    submissionStatus: PlaceOrderStatuses.Filled,
-                  }
-                : order
-            )
-          : state.localPlaceOrders,
-        localCancelOrders: hasNewFillUpdates
-          ? state.localCancelOrders.filter((order) => !filledOrderIds.includes(order.orderId))
-          : state.localCancelOrders,
       };
     },
     setFundingPayments: (state, action: PayloadAction<any>) => {
@@ -172,25 +134,6 @@ export const accountSlice = createSlice({
     },
     setTransfers: (state, action: PayloadAction<any>) => {
       state.transfers = action.payload;
-    },
-    setLatestOrder: (state, action: PayloadAction<Nullable<SubaccountOrder>>) => {
-      const { clientId, id } = action.payload ?? {};
-      state.latestOrder = action.payload;
-
-      if (clientId) {
-        state.uncommittedOrderClientIds = state.uncommittedOrderClientIds.filter(
-          (uncommittedClientId) => uncommittedClientId !== clientId
-        );
-        state.localPlaceOrders = state.localPlaceOrders.map((order) =>
-          order.clientId === clientId && order.submissionStatus < PlaceOrderStatuses.Placed
-            ? {
-                ...order,
-                orderId: id,
-                submissionStatus: PlaceOrderStatuses.Placed,
-              }
-            : order
-        );
-      }
     },
     clearOrder: (state, action: PayloadAction<string>) => ({
       ...state,
@@ -247,55 +190,10 @@ export const accountSlice = createSlice({
         state.subaccount?.orders != null &&
         payloadOrders.some((order: SubaccountOrder) => !existingOrderIds.includes(order.id));
 
-      const canceledOrderIdsInPayload = payloadOrders
-        .filter((order) => isOrderStatusCanceled(order.status))
-        .map((order) => order.id);
-
-      // ignore locally canceled orders since it's intentional and already handled
-      // by local cancel tracking and notification
-      const isOrderCanceledByBackend = (orderId: string) =>
-        canceledOrderIdsInPayload.includes(orderId) &&
-        !state.localCancelOrders.map((order) => order.orderId).includes(orderId);
-
-      const getNewCanceledOrderIds = (batch: LocalCancelAllData) => {
-        const newCanceledOrderIds = _.intersection(batch.orderIds, canceledOrderIdsInPayload);
-        return _.uniq([...(batch.canceledOrderIds ?? []), ...newCanceledOrderIds]);
-      };
-
-      const cancelUpdates = canceledOrderIdsInPayload.length
-        ? {
-            localPlaceOrders: state.localPlaceOrders.map((order) =>
-              order.orderId &&
-              canceledOrderIdsInPayload.includes(order.orderId) &&
-              isOrderCanceledByBackend(order.orderId)
-                ? {
-                    ...order,
-                    submissionStatus: PlaceOrderStatuses.Canceled,
-                  }
-                : order
-            ),
-            localCancelOrders: state.localCancelOrders.map((order) =>
-              canceledOrderIdsInPayload.includes(order.orderId)
-                ? { ...order, submissionStatus: CancelOrderStatuses.Canceled }
-                : order
-            ),
-            localCancelAlls: Object.fromEntries(
-              Object.entries(state.localCancelAlls).map(([marketId, batch]) => [
-                marketId,
-                {
-                  ...batch,
-                  canceledOrderIds: getNewCanceledOrderIds(batch),
-                },
-              ])
-            ),
-          }
-        : {};
-
       return {
         ...state,
-        subaccount: action.payload,
         hasUnseenOrderUpdates: hasNewOrderUpdates,
-        ...cancelUpdates,
+        subaccount: action.payload,
       };
     },
     setChildSubaccount: (
@@ -347,109 +245,11 @@ export const accountSlice = createSlice({
     setTradingRewards: (state, action: PayloadAction<TradingRewards>) => {
       state.tradingRewards = action.payload;
     },
-    placeOrderSubmitted: (
-      state,
-      action: PayloadAction<{ marketId: string; clientId: string; orderType: TradeTypes }>
-    ) => {
-      state.localPlaceOrders.push({
-        ...action.payload,
-        submissionStatus: PlaceOrderStatuses.Submitted,
-      });
-      state.uncommittedOrderClientIds.push(action.payload.clientId);
-    },
-    placeOrderFailed: (
-      state,
-      action: PayloadAction<{ clientId: string; errorParams: ErrorParams }>
-    ) => {
-      state.localPlaceOrders = state.localPlaceOrders.map((order) =>
-        order.clientId === action.payload.clientId
-          ? {
-              ...order,
-              errorParams: action.payload.errorParams,
-            }
-          : order
-      );
-      state.uncommittedOrderClientIds = state.uncommittedOrderClientIds.filter(
-        (id) => id !== action.payload.clientId
-      );
-    },
-    placeOrderTimeout: (state, action: PayloadAction<string>) => {
-      if (state.uncommittedOrderClientIds.includes(action.payload)) {
-        placeOrderFailed({
-          clientId: action.payload,
-          errorParams: DEFAULT_SOMETHING_WENT_WRONG_ERROR_PARAMS,
-        });
-      }
-    },
-    cancelOrderSubmitted: (state, action: PayloadAction<string>) => {
-      state.localCancelOrders.push({
-        orderId: action.payload,
-        submissionStatus: CancelOrderStatuses.Submitted,
-      });
-    },
-    cancelOrderConfirmed: (state, action: PayloadAction<string>) => {
-      state.localCancelOrders = state.localCancelOrders.map((order) =>
-        order.orderId === action.payload
-          ? { ...order, submissionStatus: CancelOrderStatuses.Canceled }
-          : order
-      );
-    },
-    cancelOrderFailed: (
-      state,
-      action: PayloadAction<{ orderId: string; errorParams: ErrorParams }>
-    ) => {
-      state.localCancelOrders = state.localCancelOrders.map((order) =>
-        order.orderId === action.payload.orderId
-          ? {
-              ...order,
-              errorParams: action.payload.errorParams,
-            }
-          : order
-      );
-    },
-    cancelAllSubmitted: (
-      state,
-      action: PayloadAction<{ marketId?: string; orderIds: string[] }>
-    ) => {
-      const { marketId, orderIds } = action.payload;
-      // when marketId is undefined, it means cancel all orders globally
-      const cancelAllKey = marketId ?? CANCEL_ALL_ORDERS_KEY;
-
-      state.localCancelAlls[cancelAllKey] = {
-        key: cancelAllKey,
-        orderIds,
-      };
-
-      state.localCancelOrders = [
-        ...state.localCancelOrders,
-        ...orderIds.map((orderId) => ({
-          orderId,
-          submissionStatus: CancelOrderStatuses.Submitted,
-          isSubmittedThroughCancelAll: true, // track this to skip certain notifications
-        })),
-      ];
-    },
-    cancelAllOrderConfirmed: (state, action: PayloadAction<string>) => {
-      const orderId = action.payload;
-      const order = state.subaccount?.orders?.toArray()?.find((o) => o.id === orderId);
-      if (!order) return;
-
-      updateCancelAllOrderIds(state, orderId, 'canceledOrderIds', order.marketId);
-      cancelOrderConfirmed(orderId);
-    },
-    cancelAllOrderFailed: (
-      state,
-      action: PayloadAction<{
-        orderId: string;
-        errorParams?: ErrorParams;
-      }>
-    ) => {
-      const { orderId, errorParams } = action.payload;
-      const order = state.subaccount?.orders?.toArray()?.find((o) => o.id === orderId);
-      if (!order) return;
-
-      updateCancelAllOrderIds(state, orderId, 'failedOrderIds', order.marketId);
-      if (errorParams) cancelOrderFailed({ orderId, errorParams });
+    clearSubaccountState: (state) => {
+      state.subaccount = undefined;
+      state.fills = undefined;
+      state.transfers = undefined;
+      state.historicalPnl = undefined;
     },
   },
 });
@@ -458,7 +258,7 @@ export const {
   setFills,
   setFundingPayments,
   setTransfers,
-  setLatestOrder,
+
   clearOrder,
   clearAllOrders,
   setOnboardingGuard,
@@ -477,43 +277,6 @@ export const {
   setTradingRewards,
   setUnbondingDelegations,
   setStakingRewards,
-  placeOrderSubmitted,
-  placeOrderFailed,
-  placeOrderTimeout,
-  cancelOrderSubmitted,
-  cancelOrderConfirmed,
-  cancelOrderFailed,
-  cancelAllSubmitted,
-  cancelAllOrderConfirmed,
-  cancelAllOrderFailed,
+
+  clearSubaccountState,
 } = accountSlice.actions;
-
-// helper functions
-const updateCancelAllOrderIds = (
-  state: AccountState,
-  orderId: string,
-  key: 'canceledOrderIds' | 'failedOrderIds',
-  cancelAllKey: string
-) => {
-  const getNewOrderIds = (cancelAll: LocalCancelAllData) => {
-    const existingOrderIds = cancelAll[key] ?? [];
-    return cancelAll.orderIds.includes(orderId) && !existingOrderIds.includes(orderId)
-      ? [...existingOrderIds, orderId]
-      : existingOrderIds;
-  };
-
-  const updateKey = (currentKey: string) => ({
-    ...state.localCancelAlls[currentKey],
-    [key]: getNewOrderIds(state.localCancelAlls[currentKey]),
-  });
-
-  state.localCancelAlls = {
-    ...state.localCancelAlls,
-    ...(state.localCancelAlls[CANCEL_ALL_ORDERS_KEY] && {
-      [CANCEL_ALL_ORDERS_KEY]: updateKey(CANCEL_ALL_ORDERS_KEY),
-    }),
-    ...(state.localCancelAlls[cancelAllKey] && {
-      [cancelAllKey]: updateKey(cancelAllKey),
-    }),
-  };
-};
